fix: directional spectrum path & ambiguity resolution, enhance wave analysis pipeline
- Fix hardcoded /home/obsis path in directional_spectrum.py (use os.path)
- Fix 180° ambiguity: class variable → instance variable with 10min timeout
- Two-pass analysis with physics-constrained QC (SWAP 4σ/4δ + steepness)
- Meter mode: IQR/MAD/jump despiking, R1-ref filter, η-domain median filter
- DFTM-CustomTRM directional spectrum with dynamic tilt correction
- Serial console: delayed analysis scheduling, STOP command support
- Config hot-reload for array_heading, filter parameters
- Add test scripts and documentation

+++ mqtt_system/services/pydiwasp/dirspec.py
@@ -131,7 +131,7 @@ def dirspec(ID, SM, EP, Options_=None):
 
     #calculate transfer parameters
     print('transfer parameters\n')
-    trm = np.empty((szd, nf, len(pidirs)))
+    trm = np.empty((szd, nf, len(pidirs)), dtype='complex128')
     kx = np.empty((szd, szd, nf, len(pidirs)))
     for m in range(szd):
         sensor_fn = _SENSOR_TYPES[ID['datatypes'][m]]
